+++ src/routes/InstancesRoute.js
@@ -7,6 +7,7 @@ import {
   filterConfig,
   renderFilters,
   segments,
+  withSearchErrors,
 } from '@folio/stripes-inventory-components';
 
 import withLocation from '../withLocation';
@@ -51,6 +52,7 @@ class InstancesRoute extends React.Component {
       storeLastSearch,
       storeLastSearchOffset,
       storeLastSegment,
+      isRequestUrlExceededLimit,
     } = this.props;
     const { segment = segments.instances } = getParams(this.props);
     const { indexes } = filterConfig[segment];
@@ -78,6 +80,7 @@ class InstancesRoute extends React.Component {
             storeLastSearch={storeLastSearch}
             storeLastSearchOffset={storeLastSearchOffset}
             storeLastSegment={storeLastSegment}
+            isRequestUrlExceededLimit={isRequestUrlExceededLimit}
           />
         )}
       </DataContext.Consumer>
@@ -89,4 +92,5 @@ export default flowRight(
   stripesConnect,
   withLocation,
   withLastSearchTerms,
+  withSearchErrors,
 )(InstancesRoute);